Rename pointer sample files and move slides
Update pointers saving the day example to be simpler. Update
relevant slide link.

Move the go for java devs slides to the correct location. Was
before lab 6, now directly after lab 5.
